discord/requirements.txt
os
datetime
discord.py
mysql-connector-python
discord-py-slash-command
random
string
pychallonge